pyproject.toml
[project]
name = "pc1600-to-json"
version = "0.1.0"
description = "Convert PC1600(x) sysex dumps to JSON and back"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mido>=1.3.3",
    "pre-commit>=4.4.0",
    "pylint>=4.0.2",
    "pysimplecli>=1.0.6",
    "pytest>=8.4.2",
    "python-rtmidi>=1.5.8",
    "ruff>=0.14.5",
]

[tool.pylint]
disable = [
    "C0114",
    "C0115",
    "C0116",
    "C0411",
    "R0913",
    "R0914",
    "R0917",
]


[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 79
indent-width = 4
lint.select = ["ALL"]
lint.ignore = [
    "ANN002",
    "ANN003",
    "COM812",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D203",
    "D211",
    "D212",
    "D213",
    "FBT001",
    "FBT002",
    "I001",
    "PLR0913",
    "PLR1722",
    "PLR2004",
    "SIM300",
    "T201",
]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-sort-within-sections = false
combine-as-imports = true
no-lines-before = ["future", "standard-library", "third-party", "first-party", "local-folder"]
